♻ [REF] Move GPX/KML handling methods to Parsers utils (refs #3947)
