+++ geotrek/trekking/parsers.py
@@ -6,27 +6,27 @@ import re
 import zipfile
 import textwrap
 from collections import defaultdict
-from datetime import date, timedelta, datetime
+from datetime import date, timedelta
 from decimal import Decimal
-from tempfile import NamedTemporaryFile
-import codecs
+
 import os
 from urllib.parse import urlparse
 
 from django.conf import settings
-from django.contrib.gis.gdal import DataSource
-from django.contrib.gis.geos import GEOSGeometry, MultiLineString, Point, LineString
+
+from django.contrib.gis.geos import GEOSGeometry, Point, LineString
 from django.utils.translation import get_language
 from django.utils.translation import gettext as _
 from paperclip.models import attachment_upload, random_suffix_regexp
 from modeltranslation.utils import build_localized_fieldname
 
-from geotrek.common.utils.file_infos import get_encoding_file
+
 from geotrek.common.models import Label, Theme, License, Attachment
 from geotrek.common.parsers import (ApidaeBaseParser, AttachmentParserMixin,
                                     GeotrekParser, GlobalImportError, Parser,
                                     RowImportError, ShapeParser, DownloadImportError,
                                     ValueImportError)
+from geotrek.common.utils.parsers import get_geom_from_gpx, get_geom_from_kml
 from geotrek.core.models import Path, Topology
 from geotrek.trekking.models import (POI, Accessibility, DifficultyLevel,
                                      OrderedTrekChild, Service, Trek,
@@ -690,13 +690,16 @@ class ApidaeTrekParser(AttachmentParserMixin, ApidaeBaseTrekkingParser):
         geom_file = self._fetch_geometry_file(plan)
 
         ext = self._get_plan_extension(plan)
-        if ext == 'gpx':
-            return ApidaeTrekParser._get_geom_from_gpx(geom_file)
-        elif ext == 'kml':
-            return ApidaeTrekParser._get_geom_from_kml(geom_file)
-        elif ext == 'kmz':
-            kml_file = zipfile.ZipFile(io.BytesIO(geom_file)).read('doc.kml')
-            return ApidaeTrekParser._get_geom_from_kml(kml_file)
+        try:
+            if ext == 'gpx':
+                return get_geom_from_gpx(geom_file)
+            elif ext == 'kml':
+                return get_geom_from_kml(geom_file)
+            elif ext == 'kmz':
+                kml_file = zipfile.ZipFile(io.BytesIO(geom_file)).read('doc.kml')
+                return get_geom_from_kml(kml_file)
+        except ValueError as e:
+            raise RowImportError(str(e))
 
     def filter_labels(self, src, val):
         typologies, environnements = val
@@ -963,115 +966,6 @@ class ApidaeTrekParser(AttachmentParserMixin, ApidaeBaseTrekkingParser):
         if url_suffix:
             return url_suffix.split('.')[1]
         return None
-
-    @staticmethod
-    def _get_geom_from_gpx(data):
-        """Given GPX data as bytes it returns a geom."""
-        # FIXME: is there another way than the temporary file? It seems not. `DataSource` really expects a filename.
-        with NamedTemporaryFile(mode='w+b', dir=settings.TMP_DIR) as ntf:
-            ntf.write(data)
-            ntf.flush()
-
-            file_path = ApidaeTrekParser._maybe_fix_encoding_to_utf8(ntf.name)
-            ds = DataSource(file_path)
-            for layer_name in ('tracks', 'routes'):
-                layer = ApidaeTrekParser._get_layer(ds, layer_name)
-                geos = ApidaeTrekParser._maybe_get_linestring_from_layer(layer)
-                if geos:
-                    break
-            else:
-                raise RowImportError("No LineString feature found in GPX layers tracks or routes")
-            geos.transform(settings.SRID)
-            return geos
-
-    @staticmethod
-    def _maybe_fix_encoding_to_utf8(file_name):
-        encoding = get_encoding_file(file_name)
-
-        # If not utf-8, convert file to utf-8
-        if encoding != "utf-8":
-            tmp_file_path = os.path.join(settings.TMP_DIR, 'fileNameTmp_' + str(datetime.now().timestamp()))
-            BLOCKSIZE = 9_048_576
-            with codecs.open(file_name, "r", encoding) as sourceFile:
-                with codecs.open(tmp_file_path, "w", "utf-8") as targetFile:
-                    while True:
-                        contents = sourceFile.read(BLOCKSIZE)
-                        if not contents:
-                            break
-                        targetFile.write(contents)
-            os.replace(tmp_file_path, file_name)
-        return file_name
-
-    @staticmethod
-    def _get_geom_from_kml(data):
-        """Given KML data as bytes it returns a geom."""
-
-        def get_geos_linestring(datasource):
-            layer = datasource[0]
-            geom = get_first_geom_with_type_in(types=['MultiLineString', 'LineString'], geoms=layer.get_geoms())
-            geom.coord_dim = 2
-            geos = geom.geos
-            if geos.geom_type == 'MultiLineString':
-                geos = geos.merged
-            return geos
-
-        def get_first_geom_with_type_in(types, geoms):
-            for g in geoms:
-                for t in types:
-                    if g.geom_type.name.startswith(t):
-                        return g
-            raise RowImportError('The attached KML geometry does not have any LineString or MultiLineString data')
-
-        with NamedTemporaryFile(mode='w+b', dir=settings.TMP_DIR) as ntf:
-            ntf.write(data)
-            ntf.flush()
-
-            file_path = ApidaeTrekParser._maybe_fix_encoding_to_utf8(ntf.name)
-            ds = DataSource(file_path)
-            geos = get_geos_linestring(ds)
-            geos.transform(settings.SRID)
-            return geos
-
-    @staticmethod
-    def _get_layer(datasource, layer_name):
-        for layer in datasource:
-            if layer.name == layer_name:
-                return layer
-
-    @staticmethod
-    def _convert_to_geos(geom):
-        # FIXME: is it right to try to correct input geometries?
-        # FIXME: how to log that info/spread errors?
-        if geom.geom_type == 'MultiLineString' and any([ls for ls in geom if ls.num_points == 1]):
-            # Handles that framework conversion fails when there are LineStrings of length 1
-            geos_mls = MultiLineString([ls.geos for ls in geom if ls.num_points > 1])
-            geos_mls.srid = geom.srid
-            return geos_mls
-
-        return geom.geos
-
-    @staticmethod
-    def _maybe_get_linestring_from_layer(layer):
-        if layer.num_feat == 0:
-            return None
-        geoms = []
-        for feat in layer:
-            if feat.geom.num_coords == 0:
-                continue
-            geos = ApidaeTrekParser._convert_to_geos(feat.geom)
-            if geos.geom_type == 'MultiLineString':
-                geos = geos.merged  # If possible we merge the MultiLineString into a LineString
-                if geos.geom_type == 'MultiLineString':
-                    raise RowImportError(_("Feature geometry cannot be converted to a single continuous LineString feature"))
-            geoms.append(geos)
-
-        full_geom = MultiLineString(geoms)
-        full_geom.srid = geoms[0].srid
-        full_geom = full_geom.merged  # If possible we merge the MultiLineString into a LineString
-        if full_geom.geom_type == 'MultiLineString':
-            raise RowImportError(_("Geometries from various features cannot be converted to a single continuous LineString feature"))
-
-        return full_geom
 
     @staticmethod
     def _find_matching_practice_in_mapping(activities_ids, mapping):
